FIX split sauce and ingredients selectors
